// packages/muya/src/state/types.ts
export interface IFootnoteBlockMeta {
    identifier: string;
}

export interface IFootnoteBlockState {
    name: 'footnote';
    meta: IFootnoteBlockMeta;
    children: TState[];
}

export type TLeafState
    = | IParagraphState
        | IAtxHeadingState
        | ISetextHeadingState
        | IThematicBreakState
        | ICodeBlockState
        | IHtmlBlockState
        | ILinkReferenceDefinitionState
        | IMathBlockState
        | IFrontmatterState
        | IDiagramState
        | ITableCellState;

export type TContainerState
    = | IBlockQuoteState
        | IOrderListState
        | IBulletListState
        | ITableState
        | ITaskListState
        | ITaskListItemState
        | IListItemState
        | ITableRowState
        | IFootnoteBlockState;

export type TState = TLeafState | TContainerState;